mgr/nfs: rename export_utils.py to ganesha_conf.py
This new name is clearer about what topic the code in the file covers.
